+++ python-phoenixdb/phoenixdb/tests/dbapi20.py
@@ -32,6 +32,7 @@ __version__ = '1.14.3'
 import unittest
 import time
 import sys
+import phoenixdb
 
 def str2bytes(sval):
     if sys.version_info < (3,0) and isinstance(sval, str):
@@ -74,7 +75,7 @@ class DatabaseAPI20Test(unittest.TestCase):
     insert = 'insert'
 
     lowerfunc = 'lower' # Name of stored procedure to convert string->lowercase
-        
+
     # Some drivers may need to override these helpers, for example adding
     # a 'commit' after the execute.
     def executeDDL1(self,cursor):
@@ -208,7 +209,7 @@ class DatabaseAPI20Test(unittest.TestCase):
                 con.rollback()
             except self.driver.NotSupportedError:
                 pass
-    
+
     def test_cursor(self):
         con = self._connect()
         try:
@@ -399,12 +400,12 @@ class DatabaseAPI20Test(unittest.TestCase):
         trouble = "thi%s :may ca%(u)se? troub:1e"
         self.assertEqual(res[0][1], trouble,
             'cursor.fetchall retrieved incorrect data, or data inserted '
-            'incorrectly. Got=%s, Expected=%s' % (repr(res[0][1]), repr(trouble)))      
+            'incorrectly. Got=%s, Expected=%s' % (repr(res[0][1]), repr(trouble)))
         self.assertEqual(res[1][1], trouble,
             'cursor.fetchall retrieved incorrect data, or data inserted '
             'incorrectly. Got=%s, Expected=%s' % (repr(res[1][1]), repr(trouble)
             ))
-        
+
     def test_executemany(self):
         con = self._connect()
         try:
@@ -575,7 +576,7 @@ class DatabaseAPI20Test(unittest.TestCase):
             self.assertEqual(len(rows),6)
             rows = [r[0] for r in rows]
             rows.sort()
-          
+
             # Make sure we get the right data back out
             for i in range(0,6):
                 self.assertEqual(rows[i],self.samples[i],
@@ -646,10 +647,10 @@ class DatabaseAPI20Test(unittest.TestCase):
                 'cursor.fetchall should return an empty list if '
                 'a select query returns no rows'
                 )
-            
+
         finally:
             con.close()
-    
+
     def test_mixedfetch(self):
         con = self._connect()
         try:
@@ -832,3 +833,17 @@ class DatabaseAPI20Test(unittest.TestCase):
         self.assertTrue(hasattr(self.driver,'ROWID'),
             'module.ROWID must be defined.'
             )
+
+    # https://issues.apache.org/jira/browse/PHOENIX-6917
+    def test_alias(self):
+        con = self._connect()
+        try:
+            cur = con.cursor(cursor_factory=phoenixdb.cursor.DictCursor)
+            self.executeDDL1(cur)
+            for sql in self._populate():
+                cur.execute(sql)
+            cur.execute('select name as beverages from %sbooze' % self.table_prefix)
+            self.assertEqual(cur.description[0][0], 'BEVERAGES')
+            self.assertEqual(cur.fetchone()['BEVERAGES'], 'Carlton Cold')
+        finally:
+            con.close()